Dado funcionando de maneiro torta ainda
Problema parcialmente resolvido
