+++ 1521730-1521816/bin/.gitignore
@@ -1,0 +1,10 @@
+/Dado1.png
+/Dado2.png
+/Dado3.png
+/Dado4.png
+/Dado5.png
+/Dado6.png
+/Main.class
+/Tabuleiro.class
+/TratadorBotao.class
+/Menu.class
