+++ 1521730-1521816/src/Main.java
@@ -19,29 +19,29 @@ public class Main extends JFrame {
 	
 	
 	//Botoes
-	static JButton b1 = new JButton("Novo Jogo");
-	static JButton b2 = new JButton("Carregar Jogo");
-	static JButton b3 = new JButton("Salvar");
-	static JButton b4 = new JButton("Lancar Dado");
+//	static JButton b1 = new JButton("Novo Jogo");
+//	static JButton b2 = new JButton("Carregar Jogo");
+//	static JButton b3 = new JButton("Salvar");
+//	static JButton b4 = new JButton("Lancar Dado");
 	
-	//Labels
-	static JLabel turno = new JLabel("À Jogar: ");
-	public static Font turno25 = new Font("Courier New", Font.BOLD, 25);
+//	//Labels
+//	static JLabel turno = new JLabel("À Jogar: ");
+//	public static Font turno25 = new Font("Courier New", Font.BOLD, 25);
 	
 	public static void main(String[] args)
 	 {
-		ConfiguraSaveGame();
-		ConfiguraLoadGame();
-		ConfiguraLancaDado();
-		ConfiguraNewGame();
-		ConfiguraLabel();
+//		ConfiguraSaveGame();
+//		ConfiguraLoadGame();
+//		ConfiguraLancaDado();
+//		ConfiguraNewGame();
+//		ConfiguraLabel();
 		ConfiguraFrame();
-		
-		frame.add(b1);
-		frame.add(b2);
-		frame.add(b3);
-		frame.add(b4);
-		frame.add(turno);
+//		
+//		frame.add(b1);
+//		frame.add(b2);
+//		frame.add(b3);
+//		frame.add(b4);
+//		frame.add(turno);
 	
 		Tabuleiro tab = new Tabuleiro();
 		
@@ -52,27 +52,27 @@ public class Main extends JFrame {
 		frame.repaint();
 	 }
 	
-	static void ConfiguraNewGame() {
-		b1.setBounds(620,50,150,50);
-	}
-	static void ConfiguraLoadGame() {
-		 b2.setBounds(620,110,150,50);
-	}
-	static void ConfiguraSaveGame() {
-		b3.setBounds(620,170,150,50); 
-		b3.setEnabled(false);
-	}
-
-	static void ConfiguraLabel() {
-		turno.setFont(turno25);
-		turno.setBounds(630, 235, 180, 30);
-	}
-
-	static void ConfiguraLancaDado() {
-		b4.setBounds(620,390,150,50);
-		b4.addActionListener(new TratadorBotao(b4));
-	
-	}
+//	static void ConfiguraNewGame() {
+//		b1.setBounds(620,50,150,50);
+//	}
+//	static void ConfiguraLoadGame() {
+//		 b2.setBounds(620,110,150,50);
+//	}
+//	static void ConfiguraSaveGame() {
+//		b3.setBounds(620,170,150,50); 
+//		b3.setEnabled(false);
+//	}
+//
+//	static void ConfiguraLabel() {
+//		turno.setFont(turno25);
+//		turno.setBounds(630, 235, 180, 30);
+//	}
+//
+//	static void ConfiguraLancaDado() {
+//		b4.setBounds(620,390,150,50);
+//		b4.addActionListener(new TratadorBotao(b4));
+//	
+//	}
 
 	static void ConfiguraFrame() {
 		frame.setSize(800, 635);
